+++ Common/src/main/java/fuzs/puzzleslib/client/core/ClientCoreServices.java
@@ -1,6 +1,5 @@
 package fuzs.puzzleslib.client.core;
 
-import fuzs.puzzleslib.client.gui.screens.Screens;
 import fuzs.puzzleslib.core.CoreServices;
 
 /**
@@ -8,11 +7,14 @@ import fuzs.puzzleslib.core.CoreServices;
  */
 public class ClientCoreServices extends CoreServices {
     /**
-     * a helper class for dealing with instances of {@link net.minecraft.client.gui.screens.Screen}
-     */
-    public static final Screens SCREENS = load(Screens.class);
-    /**
      * a collection of utility methods for registering client side content
+     *
+     * @deprecated use {@link ClientModConstructor} instead for implementing everything directly on the main mod client class
      */
+    @Deprecated(forRemoval = true)
     public static final ClientRegistration CLIENT_REGISTRATION = load(ClientRegistration.class);
+    /**
+     * important client exclusive factories
+     */
+    public static final ClientFactories FACTORIES = load(ClientFactories.class);
 }