+++ Common/src/main/java/fuzs/puzzleslib/client/core/ClientModConstructor.java
@@ -1,13 +1,17 @@
 package fuzs.puzzleslib.client.core;
 
 import fuzs.puzzleslib.client.init.builder.ModScreenConstructor;
+import fuzs.puzzleslib.client.init.builder.ModSpriteParticleRegistration;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.gui.screens.inventory.MenuAccess;
 import net.minecraft.client.gui.screens.inventory.tooltip.ClientTooltipComponent;
 import net.minecraft.client.model.geom.ModelLayerLocation;
 import net.minecraft.client.model.geom.builders.LayerDefinition;
+import net.minecraft.client.particle.ParticleProvider;
 import net.minecraft.client.renderer.blockentity.BlockEntityRendererProvider;
 import net.minecraft.client.renderer.entity.EntityRendererProvider;
+import net.minecraft.core.particles.ParticleOptions;
+import net.minecraft.core.particles.ParticleType;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EntityType;
@@ -20,44 +24,79 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import java.util.function.Function;
 import java.util.function.Supplier;
 
+/**
+ * a base class for a mods main client class, contains a bunch of methods for registering various things
+ */
 public interface ClientModConstructor {
 
+    /**
+     * runs when the mod is first constructed, on the client only really used for registering event callbacks
+     */
     default void onConstructMod() {
 
     }
 
+    /**
+     * runs after content has been registered, so it's safe to use here
+     * used to set various values and settings for already registered content
+     */
     default void onClientSetup() {
 
     }
 
+    /**
+     * @param consumer add a renderer to an entity
+     */
     default void onRegisterEntityRenderers(EntityRendererConsumer consumer) {
 
     }
 
+    /**
+     * @param consumer add a renderer to a block entity
+     */
     default void onRegisterBlockEntityRenderers(BlockEntityRendererConsumer consumer) {
 
     }
 
+    /**
+     * @param consumer add a client tooltip component to a common tooltip component
+     */
     default void onRegisterClientTooltipComponents(ClientTooltipComponentConsumer consumer) {
 
     }
 
-    default void onRegisterParticleProviders() {
+    /**
+     * @param consumer add particle providers for a particle type
+     */
+    default void onRegisterParticleProviders(ParticleProviderConsumer consumer) {
 
     }
 
+    /**
+     * @param consumer register a screen for a menu type
+     */
     default void onRegisterMenuScreens(MenuScreenConsumer consumer) {
 
     }
 
+    /**
+     * @param consumer add a sprite to a texture atlas
+     */
     default void onRegisterAtlasSprites(AtlasSpriteConsumer consumer) {
 
     }
 
+    /**
+     * @param consumer add a layer definition for a {@link ModelLayerLocation}
+     */
     default void onRegisterLayerDefinitions(LayerDefinitionConsumer consumer) {
 
     }
 
+    /**
+     * register a renderer for an entity
+     */
+    @FunctionalInterface
     interface EntityRendererConsumer {
 
         /**
@@ -67,9 +106,13 @@ public interface ClientModConstructor {
          * @param entityRendererProvider entity renderer provider
          * @param <T> type of entity
          */
-        <T extends Entity> void registerEntityRenderer(EntityType<? extends T> entityType, EntityRendererProvider<T> entityRendererProvider);
+        <T extends Entity> void register(EntityType<? extends T> entityType, EntityRendererProvider<T> entityRendererProvider);
     }
 
+    /**
+     * register a renderer for a block entity
+     */
+    @FunctionalInterface
     interface BlockEntityRendererConsumer {
 
         /**
@@ -79,9 +122,13 @@ public interface ClientModConstructor {
          * @param blockEntityRendererProvider   block entity renderer provider
          * @param <T> type of entity
          */
-        <T extends BlockEntity> void registerBlockEntityRenderer(BlockEntityType<? extends T> blockEntityType, BlockEntityRendererProvider<T> blockEntityRendererProvider);
+        <T extends BlockEntity> void register(BlockEntityType<? extends T> blockEntityType, BlockEntityRendererProvider<T> blockEntityRendererProvider);
     }
 
+    /**
+     * register a client-side tooltip component factory
+     */
+    @FunctionalInterface
     interface ClientTooltipComponentConsumer {
 
         /**
@@ -91,9 +138,37 @@ public interface ClientModConstructor {
          * @param factory factory for creating {@link ClientTooltipComponent} from <code>type</code>
          * @param <T>     type of common component
          */
-        <T extends TooltipComponent> void registerClientTooltipComponent(Class<T> type, Function<? super T, ? extends ClientTooltipComponent> factory);
+        <T extends TooltipComponent> void register(Class<T> type, Function<? super T, ? extends ClientTooltipComponent> factory);
     }
 
+    /**
+     * register a particle provider for a particle type
+     */
+    interface ParticleProviderConsumer {
+
+        /**
+         * registers a factory for a particle type client side
+         *
+         * @param type     particle type (registered separately)
+         * @param provider particle factory
+         * @param <T>      type of particle
+         */
+        <T extends ParticleOptions> void register(ParticleType<T> type, ParticleProvider<T> provider);
+
+        /**
+         * registers a factory for a particle type client side
+         *
+         * @param type     particle type (registered separately)
+         * @param factory particle factory
+         * @param <T>      type of particle
+         */
+        <T extends ParticleOptions> void register(ParticleType<T> type, ModSpriteParticleRegistration<T> factory);
+    }
+
+    /**
+     * register a screen for a menu type
+     */
+    @FunctionalInterface
     interface MenuScreenConsumer {
 
         /**
@@ -108,6 +183,10 @@ public interface ClientModConstructor {
         <M extends AbstractContainerMenu, U extends Screen & MenuAccess<M>> void registerMenuScreen(MenuType<? extends M> menuType, ModScreenConstructor<M, U> factory);
     }
 
+    /**
+     * stitch a custom sprite onto an atlas
+     */
+    @FunctionalInterface
     interface AtlasSpriteConsumer {
 
         /**
@@ -116,9 +195,13 @@ public interface ClientModConstructor {
          * @param atlasId the atlas to register to, since 1.14 there are multiples
          * @param spriteId the sprite to register
          */
-        void registerAtlasSprite(ResourceLocation atlasId, ResourceLocation spriteId);
+        void register(ResourceLocation atlasId, ResourceLocation spriteId);
     }
 
+    /**
+     * register layer definitions for entity models
+     */
+    @FunctionalInterface
     interface LayerDefinitionConsumer {
 
         /**
@@ -127,6 +210,6 @@ public interface ClientModConstructor {
          * @param layerLocation model location
          * @param supplier      layer definition supplier
          */
-        void registerLayerDefinition(ModelLayerLocation layerLocation, Supplier<LayerDefinition> supplier);
+        void register(ModelLayerLocation layerLocation, Supplier<LayerDefinition> supplier);
     }
 }